services/sentiment-service/requirements.txt
kafka-python-ng==2.2.2
transformers==4.44.2
torch==2.4.1
redis==5.0.8
flask==3.0.3
python-dotenv==1.0.1
